using ALTER syntax for MySQL index creation
This addresses an issue in MySQL[1] where adding an index via CTEATE will
corrupt the casing of the table name. For example:

    CREATE TABLE `Example` (`bar` VARCHAR(1));
    CREATE INDEX `foo` ON `Example` (`bar`);

The table name after these queries will be lowercased, "example."

In my.cnf, setting `lower_case_table_names=0` does address the problem,
however, it is strongly advised against[2] under cerain circumstances.
Mac OSX in particular prefers the option to be set at `2`, which is the
underlying cause of the problem.

[1]: http://bugs.mysql.com/bug.php?id=48875
[2]: http://dev.mysql.com/doc/refman/5.0/en/identifier-case-sensitivity.html