Create migrations table in specified schema
In postgres, if you create a migrations table in one schema of your database,
and then try to use this for migrations in another schema, it will not create
the migrations table in the second schema which will prevent any migrations
from running on that schema. This creates the migrations table in that second
schema when the schema parameter is specified.