Extend startup with logging and failure handling on 0.9.x versions
Before this patch users may ran into a local installation older than 0.9.x, this patch
ensures that the version is compatible to be preffered over the global version. If the
local version is to old (<0.10.0) the global version will be used instead. If this is
the case a warn message gets displayed to the user.

Additionally added informational logging within the verbose mode if a local version gets
preferred.

solves #317