Avoid warning in promise/callback bridge code
When Bluebird detects when a promise is created within a promise
handler, but is not returned, it [emits a warning][warn] 'a promise was
created in a handler but was not returned from it'.

This happens if the `callback()` function is implemented using Promises,
but uses `nodeify()` to make it callback compatible. By using nodeify
here, it handles the returned Promise (if any) and avoids the warning.

 [warn]: https://github.com/petkaantonov/bluebird/blob/master/docs/docs/warning-explanations.md