refactor(globals): remove several globals, this might break things
There are several things that are exposed as global. For example Migrator, Seeder, Class, or the async and dbm globals.

Most probably finally removing async and dbm which have been marked deprecated, will end up for some people in a breaking change. But we need to remove all those
globals, to not interfere with projects using db-migrate. DB-Migrate is not just a cli module anymore though. In this step we now also finally removed internals completely
from the API, and needed to adjust the onComplete callback to actually do this. This change could also lead to a breaking change and should thus be communicated

All in all this is a big clean up and also two not necessary functions were removed.