refactor(transition): remove transitioner entirely
since it is not needed anymore.

BREAKING CHANGE: the transitioner will disappear from the API
entirely. The need for it disappeared since it was there to
help with the migration from very old migration schemas to
the new ones that did not support very old globals and async
the library provided by db-migrate itself.
Users that for some reason need that can get it from the v0.11.x
versions and then migrate to the newest version afterwards.

Refers to #627

Signed-off-by: Tobias Gurtzick <magic@wizardtales.com>